changing the type of the varray and improving the function

+++ include/varArray.h
@@ -4,12 +4,12 @@
 typedef struct
 {
     int size;
-    int * varArr;
+    double * varArr;
 }varray;
 
 varray varArray_cre( int init_size );
-int varArray_get( varray * a, int index );
-void varArray_assign( varray * a , int index , int number );
+double varArray_get( varray * a, int index );
+void varArray_assign( varray * a , int index , double number );
 int varArray_size( varray * a );
 void varArray_inflate( varray *a , int more_size );
 void varArray_free( varray *a );
